pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-development-orchestrator"
version = "0.1.0"
description = "Universal AI-powered development automation platform"
authors = ["Development Team <dev@company.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/company/ai-development-orchestrator"
repository = "https://github.com/company/ai-development-orchestrator"
documentation = "https://docs.company.com/ai-orchestrator"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "core"}, {include = "plugins"}, {include = "workflows"}]

[tool.poetry.scripts]
ai-orchestrator = "workflows.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
redis = "^5.0.0"
anthropic = "^0.34.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
httpx = "^0.25.0"
GitPython = "^3.1.40"
Jinja2 = "^3.1.2"
PyYAML = "^6.0.1"
python-dotenv = "^1.0.0"
PyJWT = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# cryptography = ">=3.0.0"
structlog = "^23.2.0"
prometheus-client = "^0.19.0"
psutil = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.7.1"
bandit = "^1.7.5"

[tool.poetry.extras]
celery = ["celery"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["core", "plugins", "agents", "monitoring"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "performance: marks tests as performance tests",
    "live_service: marks tests that require live external services"
]

[tool.coverage.run]
source = ["core", "agents", "plugins"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.flake8]
max-line-length = 120
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
    "F401",  # module imported but unused
    "F541",  # f-string is missing placeholders
    "F811",  # redefinition of unused variable
    "F841",  # local variable assigned but never used
    "E402",  # module level import not at top of file (needed for plugin imports)
    "W293",  # blank line contains whitespace
    "W291",  # trailing whitespace
]